Create a re usable history page GUI template (#870)
* feat(2): Refactored History page to a package

* chore(2): Replaced file with actual package

* chore(2): Changed Description

* chore(2): Synced ended chats sql with backoffice

* chore(2): Bumped Header Version

* chore(2): Remove Unused components

* chore(2): Removed Used component from index

* chore(2): Bumped Package Version
