+++ DSL/Ruuter.private/POST/chats/get.yml
@@ -1,0 +1,29 @@
+declaration:
+  call: declare
+  version: 0.1
+  description: "Decription placeholder for 'POST'"
+  method: post
+  accepts: json
+  returns: json
+  namespace: backoffice
+  allowlist:
+    body:
+      - field: chatId
+        type: string
+        description: "Body field 'chatId'"
+
+extractRequestData:
+  assign:
+    chatId: ${incoming.body.chatId}
+
+getChatById:
+  call: http.post
+  args:
+    url: "[#TRAINING_USERS_RESQL]/get-chat-by-id"
+    body:
+      id: ${chatId}
+  result: res
+
+return_result:
+   return: ${res.response.body[0]}  
+   next: end
