+++ GUI/package.json
@@ -14,6 +14,7 @@
     "@buerokratt-ria/header": "^0.1.21",
     "@buerokratt-ria/menu": "^0.2.6",
     "@buerokratt-ria/styles": "^0.0.1",
+    "@buerokratt-ria/common-gui-components": "^0.0.3",
     "@fontsource/roboto": "^4.5.8",
     "@formkit/auto-animate": "^1.0.0-beta.5",
     "@radix-ui/react-accessible-icon": "^1.0.1",
@@ -54,6 +55,7 @@
     "reactflow": "^11.4.0",
     "regexify-string": "^1.0.19",
     "rxjs": "^7.8.1",
+    "rich-textarea": "^0.26.4",
     "use-debounce": "^10.0.1",
     "usehooks-ts": "^2.9.1",
     "zustand": "^4.3.2"
