--- GUI/src/components/HistoricalChat/HistoricalChat.scss
@@ -1,271 +1,0 @@
-@import 'src/styles/tools/spacing';
-@import 'src/styles/tools/color';
-@import 'src/styles/settings/variables/other';
-@import 'src/styles/settings/variables/typography';
-
-.historical-chat {
-  $self: &;
-  display: flex;
-  justify-content: flex-end;
-  position: relative;
-  height: 100%;
-  white-space: pre-wrap;
-
-  &__body {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-  }
-
-  &__header {
-    padding: get-spacing(haapsalu);
-    background-color: get-color(extra-light);
-    border-bottom: 1px solid get-color(black-coral-2);
-  }
-
-  &__group-wrapper {
-    flex: 1;
-    padding: get-spacing(haapsalu) get-spacing(haapsalu) 0;
-    overflow: auto;
-    overflow-anchor: none;
-
-    #anchor {
-      overflow-anchor: auto;
-      height: 1px;
-      margin-top: 16px;
-    }
-  }
-
-  &__toolbar-row {
-    border-top: 1px solid get-color(black-coral-2);
-    padding: get-spacing(paldiski) get-spacing(haapsalu);
-  }
-
-  &__group {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    padding-left: 56px;
-    position: relative;
-
-    &:not(:last-child) {
-      margin-bottom: 8px;
-    }
-
-    &--event {
-      #{$self} {
-        &__event-message {
-          display: flex;
-          position: relative;
-          isolation: isolate;
-          padding: 6px 0;
-
-          p {
-            color: get-color(black-coral-12);
-            font-size: $veera-font-size-80;
-            line-height: $veera-line-height-500;
-            padding: 4px 8px 4px 0;
-            background-color: get-color(white);
-          }
-
-          &::after {
-            content: '';
-            display: block;
-            height: 1px;
-            background-color: get-color(black-coral-2);
-            position: absolute;
-            top: 50%;
-            left: 56px;
-            right: 0;
-            transform: translateY(-50%);
-            z-index: -1;
-          }
-        }
-      }
-    }
-
-    &--buerokratt,
-    &--chatbot {
-      #{$self} {
-        &__group-initials {
-          background-color: get-color(sapphire-blue-10);
-          left: 0;
-          top: 0;
-        }
-
-        &__group-name {
-          color: get-color(sapphire-blue-10);
-          text-align: left;
-          margin-left: 0;
-        }
-
-        &__messages {
-          display: flex;
-          flex-direction: column;
-          gap: 4px;
-          align-items: flex-start;
-        }
-
-        &__message-text {
-          background-color: get-color(sapphire-blue-10);
-          color: get-color(white);
-          margin-left: 0;
-
-          :any-link {
-            color:  get-color(white);
-          }
-
-          &:hover {
-            background-color: get-color(sapphire-blue-13);
-          }
-        }
-      }
-    }
-
-    &--backoffice-user {
-      #{$self} {
-        &__group-initials {
-          background-color: get-color(black-coral-10);
-          color: get-color(white);
-          left: 0;
-        }
-
-        &__group-name {
-          text-align: left;
-          margin-left: 0;
-        }
-
-        &__messages {
-          display: flex;
-          flex-direction: column;
-          gap: 4px;
-          align-items: flex-start;
-        }
-
-        &__message-text {
-          background-color: get-color(black-coral-10);
-          color: get-color(white);
-          margin-left: 0;
-
-          :any-link {
-            color:  get-color(white);
-          }
-
-          &:hover {
-            background-color: #535665;
-          }
-        }
-      }
-    }
-  }
-
-  &__group-initials {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 40px;
-    height: 40px;
-    background-color: get-color(black-coral-1);
-    border-radius: 50%;
-    font-size: $veera-font-size-80;
-    line-height: $veera-line-height-500;
-    font-weight: $veera-font-weight-delta;
-    color: get-color(black-coral-6);
-    position: absolute;
-    right: 0;
-    top: 0;
-  }
-
-  &__group-name {
-    font-size: $veera-font-size-80;
-    line-height: $veera-line-height-500;
-    font-weight: $veera-font-weight-delta;
-    text-transform: capitalize;
-    margin-left: 10rem;
-    margin-right: 3rem;
-    text-align: right;
-    padding-top: 7px;
-  }
-
-  &__messages {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    align-items: flex-end;
-  }
-
-  &__message {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  }
-
-  &__message-text {
-    display: flex;
-    align-items: center;
-    padding: 4px 8px;
-    background-color: get-color(black-coral-0);
-    border-radius: 8px 8px 8px 4px;
-    font-size: $veera-font-size-100;
-    line-height: $veera-line-height-500;
-    transition: all .25s ease-out;
-    cursor: pointer;
-
-    :any-link {
-      text-decoration:  underline;
-    }
-
-    &:hover {
-      background-color: get-color(black-coral-1);
-    }
-  }
-
-  &__message-date {
-    color: get-color(black-coral-6);
-    font-size: 11px;
-    line-height: 20px;
-  }
-}
-
-.reset {
-  margin: 0;
-  padding: 0;
-  line-height: 1.3;
-}
-
-.reset h1,
-.reset h2,
-.reset h3,
-.reset h4,
-.reset h5,
-.reset h6 {
-  margin: 0;
-  padding: 0;
-}
-
-.reset p {
-  margin:  0 0 0.5rem;
-  padding: 0;
-}
-
-.reset ul,
-.reset ol {
-  margin: 0 0 1rem;
-  padding-left: 1.5rem;
-}
-
-.reset li {
-  margin: 0;
-  padding: 0;
-}
-
-.reset a {
-  text-decoration: none;
-  color: inherit;
-}
