+++ GUI/src/types/intent.ts
@@ -1,5 +1,6 @@
 export interface Intent {
   id: string;
+  intent?: string
   description: string | null;
   inModel: boolean;
   modifiedAt: string;
