+++ GUI/src/utils/get-chat-event-message.ts
@@ -38,18 +38,16 @@ export const getChatEventMessage = (message: Message) => {
     case CHAT_EVENTS.ASSIGN_PENDING_CHAT_CSA:
     case CHAT_EVENTS.PENDING_USER_REACHED:
     case CHAT_EVENTS.PENDING_USER_NOT_REACHED:
-      return i18n.t(`chat.event.${event?.toLowerCase()}`, {
+      return i18n.t(`chat.events.${event?.toLowerCase()}`, {
         name: message.authorFirstName ?? 'CSA',
       });
     case CHAT_EVENTS.USER_AUTHENTICATED:
-      return i18n.t(`chat.event.${event?.toLowerCase()}`, {
-        name: `${message.authorFirstName ?? ''} ${
-          message.authorLastName ?? ''
-        }`,
+      return i18n.t(`chat.events.${event?.toLowerCase()}`, {
+        name: `${message.authorFirstName ?? ''} ${message.authorLastName ?? ''}`,
         date: format(new Date(message.authorTimestamp), 'dd.MM.yyyy HH:mm:ss'),
       });
     default:
-      return i18n.t(`chat.event.${event?.toLowerCase()}`, {
+      return i18n.t(`chat.events.${event?.toLowerCase()}`, {
         date: format(new Date(authorTimestamp), 'dd.MM.yyyy HH:mm:ss'),
       });
   }
